Fix tests
- Move them to the correct class. They were exercising filter matching,
  not parsing.
- Remove a bunch of unreadable tuple stuff that was mostly obfuscating
  the readability without actually improving test coverage.
- Make tests fail everywhere rather than on CI only. They were failing
  because they were written in a way that was implicitly dependent on
  the local computer's timezone.
